add python wrapper and auto-tests with python
It covers the following topics:
1. Now (well, after publishing to PyPi) db_verifier can be installed into a Python env in one line and can be started as a standalone command.
2. As a standalone command it can be used in any CI pipeline to verify a DB, print a report in CSV/TXT/JSON and return the proper OS error code.
3. As a Python function it can be imported in a Python project and become a part of Python auto test process.
For example, it lays out on Django testing facilities very good:
Django initializes an empty test DB from its models and migrations.
Next a connection to this test DB is provided to db_verifier which is executed in a usual Pytest and raises an exception if there are errors.
4. There are now auto-tests for the current SQL test files - load SQL to a temp DB, compare what is returned by the db_verifier with the expected set of values.